pipeline/requirements.txt
google-cloud-pubsub==2.13.6
google-api-core==2.8.2
google-apitools==0.5.32
plotly==5.10.0
itables==1.2.0
xgboost==1.6.2
google-cloud-bigquery==3.5.0
plotly==5.10.0
google-cloud-pipeline-components
google-cloud-aiplatform==1.22.0
google-cloud-bigquery
kfp
pandas
numpy
